feature: modulize admin job
