+++ src/admin/actions/acceptGroupApply.js
@@ -1,6 +1,7 @@
 import { Group, GroupApply, User } from "../../db";
 import { logger } from "../../utils/logger";
 import { sendApplyDoneMessage } from "../../utils/slack";
+import acceptGroup from "../api/acceptGroup";
 
 export const acceptGroupAction = {
   actionType: "record",
@@ -8,37 +9,9 @@ export const acceptGroupAction = {
   handler: async (req, res, context) => {
     const { record } = context;
     const currentAdmin = req.adminUser;
-
     const groupName = record.params.name;
-    const adminName = currentAdmin.user.username;
 
-    logger.admin.info(
-      "Accept Group; name: %s, adminUser: %s",
-      groupName,
-      adminName,
-    );
-
-    const newGroup = await GroupApply.findOneAndDelete({ name: groupName });
-    const newGroupJSON = newGroup.toJSON({ virtuals: false });
-    delete newGroupJSON._id;
-    newGroupJSON.members.forEach((member, i) => {
-      delete newGroupJSON.members[i]._id;
-    });
-    const created = await Group.create(newGroupJSON);
-
-    currentAdmin.actionHistory.push({
-      name: "acceptGroup",
-      target: created._id,
-    });
-
-    await Promise.all([
-      ...newGroupJSON.members.map(({ user }) =>
-        User.findByIdAndUpdate(user._id, { $push: { groups: created._id } }),
-      ),
-      currentAdmin.save(),
-    ]);
-
-    sendApplyDoneMessage(groupName, currentAdmin);
+    const created = await acceptGroup(currentAdmin, groupName);
 
     return {
       record: record.toJSON(created),
